Set VERSION in Makefile correctly
For `VERSION=<foo> make artifacthub-pkg.yml`, we expect for VERSION
to come from the tag we are creating.

For `VERSION=<foo> make annotated-policy.yml`, we expect for VERSION
to be already set in the artifacthub-pkg.yml file under `version: foo`,
and artifacthub-pkg.yml must already exist per the target definition.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>